Show the list of available symbols underneath the text entry box
# This required moving the parse function out from inside the modal, since
# we also need to use it to sanitise the displayed list. Since we also now
# use it outside, we need to remove `_no_alphabet` too.
