+++ src/app/components/elements/modals/InequalityModal.tsx
@@ -4,6 +4,7 @@ import {Inequality, makeInequality, WidgetSpec} from "inequality";
 import katex from "katex";
 import _uniqWith from 'lodash/uniqWith';
 import _isEqual from 'lodash/isEqual';
+import {parsePseudoSymbolicAvailableSymbols} from "../../../services/questions";
 
 class MenuItem {
     public type: string;
@@ -111,7 +112,7 @@ export class InequalityModal extends React.Component<InequalityModalProps> {
     public constructor(props: InequalityModalProps) {
         super(props);
 
-        this._availableSymbols = Array.from(new Set(this.parsePseudoSymbols(props.availableSymbols)));
+        this._availableSymbols = Array.from(new Set(parsePseudoSymbolicAvailableSymbols(props.availableSymbols)));
 
         this.state = {
             sketch: props.sketch as Inequality,
@@ -270,6 +271,7 @@ export class InequalityModal extends React.Component<InequalityModalProps> {
                         customMenuItems.otherFunctions.push(this.makeMathsLogFunctionItem(functionName));
                     } else {
                         // What
+                        // eslint-disable-next-line no-console
                         console.warn(`Could not parse available symbol "${availableSymbol} as a function"`);
                     }
                 } else if (availableSymbol.startsWith('Derivative')) {
@@ -396,41 +398,6 @@ export class InequalityModal extends React.Component<InequalityModalProps> {
         document.body.style.height = '';
         document.body.style.overflow = '';
         document.body.style.touchAction = 'auto';
-    }
-
-    private parsePseudoSymbols(availableSymbols?: string[]) {
-        if (!availableSymbols) return;
-        let theseSymbols = availableSymbols.slice(0).map(s => s.trim());
-        let i = 0;
-
-        // Note: _no_alphabet is already taken care of in the constructor.
-        while (i < theseSymbols.length) {
-            if (theseSymbols[i] === '_trigs') {
-                theseSymbols.splice(i, 1, 'cos()', 'sin()', 'tan()');
-                i += 3;
-            } else if (theseSymbols[i] === '_1/trigs') {
-                theseSymbols.splice(i, 1, 'cosec()', 'sec()', 'cot()');
-                i += 3;
-            } else if (theseSymbols[i] === '_inv_trigs') {
-                theseSymbols.splice(i, 1, 'arccos()', 'arcsin()', 'arctan()');
-                i += 3;
-            } else if (theseSymbols[i] === '_inv_1/trigs') {
-                theseSymbols.splice(i, 1, 'arccosec()', 'arcsec()', 'arccot()');
-                i += 3;
-            } else if (theseSymbols[i] === '_hyp_trigs') {
-                theseSymbols.splice(i, 1, 'cosh()', 'sinh()', 'tanh()', 'cosech()', 'sech()', 'coth()');
-                i += 6;
-            } else if (theseSymbols[i] === '_inv_hyp_trigs') {
-                theseSymbols.splice(i, 1, 'arccosh()', 'arcsinh()', 'arctanh()', 'arccosech()', 'arcsech()', 'arccoth()');
-                i += 6;
-            } else if (theseSymbols[i] === '_logs') {
-                theseSymbols.splice(i, 1, 'log()', 'ln()');
-                i += 2;
-            } else {
-                i += 1;
-            }
-        }
-        return theseSymbols;
     }
 
     private convertToLatexIfGreek(s: string): string {
