+++ src/app/services/questions.ts
@@ -34,3 +34,38 @@ export const HUMAN_QUESTION_TYPES = new Map([
     ["isaacSymbolicLogicQuestion", "Boolean logic"],
     ["default", "Multiple choice"]
 ]);
+
+export const parsePseudoSymbolicAvailableSymbols = (availableSymbols?: string[]) => {
+    if (!availableSymbols) return;
+    let theseSymbols = availableSymbols.slice(0).map(s => s.trim());
+    let i = 0;
+    while (i < theseSymbols.length) {
+        if (theseSymbols[i] === '_trigs') {
+            theseSymbols.splice(i, 1, 'cos()', 'sin()', 'tan()');
+            i += 3;
+        } else if (theseSymbols[i] === '_1/trigs') {
+            theseSymbols.splice(i, 1, 'cosec()', 'sec()', 'cot()');
+            i += 3;
+        } else if (theseSymbols[i] === '_inv_trigs') {
+            theseSymbols.splice(i, 1, 'arccos()', 'arcsin()', 'arctan()');
+            i += 3;
+        } else if (theseSymbols[i] === '_inv_1/trigs') {
+            theseSymbols.splice(i, 1, 'arccosec()', 'arcsec()', 'arccot()');
+            i += 3;
+        } else if (theseSymbols[i] === '_hyp_trigs') {
+            theseSymbols.splice(i, 1, 'cosh()', 'sinh()', 'tanh()', 'cosech()', 'sech()', 'coth()');
+            i += 6;
+        } else if (theseSymbols[i] === '_inv_hyp_trigs') {
+            theseSymbols.splice(i, 1, 'arccosh()', 'arcsinh()', 'arctanh()', 'arccosech()', 'arcsech()', 'arccoth()');
+            i += 6;
+        } else if (theseSymbols[i] === '_logs') {
+            theseSymbols.splice(i, 1, 'log()', 'ln()');
+            i += 2;
+        } else if (theseSymbols[i] === '_no_alphabet') {
+            theseSymbols.splice(i, 1);
+        } else {
+            i += 1;
+        }
+    }
+    return theseSymbols;
+}
