+++ src/scss/common/questions.scss
@@ -126,6 +126,20 @@ button {
   .eqn-editor-help {
     min-width: 0;
   }
+
+  .eqn-editor-symbols {
+    font-size: 0.75rem;
+    margin: 10px 0;
+
+    & pre {
+      font-size: 0.75rem;
+      display: inline;
+      white-space: pre-wrap;
+      background: inherit;
+      border: 0;
+      padding: 0;
+    }
+  }
 }
 
 @include media-breakpoint-up(md) {
